feat: refactor services to use Feign clients for external API communication and improve code structure

+++ services/tracking-service/src/main/java/pe/edu/upc/tracking_service/tracking/domain/services/TrackingCommandService.java
@@ -1,16 +1,19 @@
 package pe.edu.upc.tracking_service.tracking.domain.services;
 
-import pe.edu.upc.center.jameoFit.tracking.domain.model.aggregates.Tracking;
-import pe.edu.upc.center.jameoFit.tracking.domain.model.commands.CreateMealPlanEntryToTrackingCommand;
-import pe.edu.upc.center.jameoFit.tracking.domain.model.commands.CreateTrackingCommand;
-import pe.edu.upc.center.jameoFit.tracking.domain.model.commands.RemoveMealPlanEntryFromTrackingCommand;
-import pe.edu.upc.center.jameoFit.tracking.domain.model.commands.UpdateMealPlanEntryInTrackingCommand;
+import pe.edu.upc.tracking_service.tracking.domain.model.aggregates.Tracking;
+import pe.edu.upc.tracking_service.tracking.domain.model.commands.CreateMealPlanEntryToTrackingCommand;
+import pe.edu.upc.tracking_service.tracking.domain.model.commands.CreateTrackingCommand;
+import pe.edu.upc.tracking_service.tracking.domain.model.commands.RemoveMealPlanEntryFromTrackingCommand;
+import pe.edu.upc.tracking_service.tracking.domain.model.commands.UpdateMealPlanEntryInTrackingCommand;
 
 import java.util.Optional;
 
 public interface TrackingCommandService {
-    int handle(CreateTrackingCommand command);
-    int handle(CreateMealPlanEntryToTrackingCommand command);
+    Long handle(CreateTrackingCommand command);
+
+    Long handle(CreateMealPlanEntryToTrackingCommand command);
+
     void handle(RemoveMealPlanEntryFromTrackingCommand command);
+
     Optional<Tracking> handle(UpdateMealPlanEntryInTrackingCommand command);
 }
